29454 [Timeline] Migrate UTs to ES2015 modules
